requirements.txt
pandas
openpyxl
psycopg2-binary
mysql-connector-python
